[SYCL] Add max work-group size kernel properties
This patch adds two kernel properties to allow users to specify the
maximum work-group size that a kernel will be invoked with.

The `max_work_group_size` property corresponds to the
`intel::max_work_group_size` function attribute, but can be specified
with 1, 2, or 3 dimensions (unlike the attribute which accepts only 3).

The `max_total_work_group_size` property is similar but is always a
single value which denotes the combined total work-group size. This can
be used when the user cannot guarantee a maximum bound in each of the
dimensions they wish to run the kernel, but can guarantee a total. This
acts similarly to CUDA's `maxThreadsPerBlock` launch bounds property.

This patch also wires up the 'max_work_group_size' property to the
equivalent SPIR-V execution mode, which should hopefully improve certain
use cases.