[SYCL][NFCI][ABI-Break] Move handler members to impl
This moves some of the members in the handler class to its impl class.
Doing so allows developers to change these arguments without breaking
ABI. This also moves more implementation details, such as command-group
classes, launch configuration information, argument information and
HostTask tracking, into sources to avoid hard-to-find ABI breaks in the
communication between headers and runtime library.

In addition to this, the following improvements are made:
* The HostKernel class has been simplified to no longer have call and
  runOnHost functions.
* The HostTask wrapper class has been moved to sources and the owner
  has been changed from a unique_ptr to a shared_ptr, which prevents the
  need for including host_task_impl.hpp in odd places.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>