Allow llvm-spirv to be built natively in LLVM cross builds. (#2633)
When llvm-spirv is built in-tree as part of LLVM, and the LLVM build
also includes libclc, libclc will require a native version of llvm-spirv
to be built. To accommodate this, call setup_host_tool. This change by
itself is intended to have no effect at the moment, but enables a future
LLVM change.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/a276363ccac20bf