[SYCL] Remove `sycl::feature_not_supported` exception (#14423)
One of the SYCL 1.2 exception subclasses being removed during the ABI
breaking window since these subclasses have been removed in SYCL 2020.

We still want to be able to carry information about backend error code,
so `MPIErr` data member in `sycl::exception` remains. However, since
none of the standard ctors can accept it (per the specification), I've
introduced a new helper `detail::set_pi_error` to set it. It should only
be used when propagating an error comming from the PI/UR level and *NOT*
for the error conditions originated in the SYCL RT (which is a change
from previous implementation but a logical/justifiable one).