Refactor L0 command-buffer cleanup.
Coverity has reported that there is the possibility of the
command-buffer destructor throwing an exception from its calls
to `CleanupCompletedEvent`.

This is from an underlying `throw` in
[GetQueue](https://github.com/oneapi-src/unified-runtime/blob/main/source/adapters/level_zero/queue.hpp#L890)
when a `dynamic_cast` doesn't behave as expected. I initially tried
changing this to an assert but the V2 Level Zero adapter is triggering
this case while in it's developmental state, so an assert brings down the
whole test suite rather than just being reported as a fail (the UR
loader catches exceptions and returns them as error codes).

Therefore this PR addresses the issue by moving the calls to
`CleanupCompletedEvent()` outside of the command-buffer destructor. This allows
us us to report an error from command-buffer release by doing cleanup prior to
calling the destructor, without the risk of throwing an exception.