[SYCL][E2E] Fix some tests in multi-device mode (#13510)
The atomic tests can be run by multiple SYCL devices, in which case the
compiler needs to be told which device the `--cuda-gpu-arch` parameter
applies to. This is done in the same vein as other occurrences of this
parameter.

The bfloat16 tests fail on some CPU devices due to a known issue with
opaque pointers.