[OpenCL] Set WG dimension metadata for all reqd_work_group_size attrs
This makes the OpenCL and SYCL function attributes lower consistently to
the same LLVM IR. This could be useful, e.g., in a middle-end or backend
pass processing the `!reqd_work_group_size` metadata that has no
knowledge of the front-end attributes. It might not even be able to
distinguish between OpenCL and SYCL kernels since they usually share the
same calling conventions, metadata, etc. The choices the pass would have
to make to work around the possible presence or absence of the
!work_group_num_dim metadata and the assumptions either would imply
would be greatly simplified with a uniform metadata format.

This also cleans up the SYCL side of the attribute handling somewhat.