[ASAN] fix recursive initialization of sanitizer context
The AsanOptions class, a singleton, was being instantiated inside
of the sanitizer context constructor, but, inside of its contructor,
it was using the logger from inside of the context. This only worked
before because the context variable was a global and the logger just
so happened to be initialized prior to the AsanOptions.

Now that the layer contexts are not globals, this was causing
a deadlock on trying to retrieve a context while it was being
initialized. The fix is to pass the logger manually to the
AsanOptions class.

A better fix might be to refactor AsanOptions to no longer be a static,
but I was trying to minimize changes since this is blocking intel/llvm
update.