[SYCL][NVPTX] Emit reqd_work_group_size attributes as NVVM annotations (#14502)
Only emit the provided values as annotations in the LLVM IR. The NVPTX
backend will pad missing values with 1s. This suits the fact that the
attribute must provide as many values as the dimensionality of the
work-group, and we can assume that the work-group size of unused
dimensions is 1.