[ClangOffloadWrapper] Disable note section verification for now
This is exposed by

| * fb5a38bb4930736b0aab3ce428b60245921f982f 2024-07-11 [llvm-objcopy]
Add verification of added .note section format sguelton@mozilla.com
Somehow the ELFNotes added by ClangOffloadWrapper.cpp does NOT pass the
verification.

Looks like we used the "--add-section" wrongly – we call llvm-objcopy
ONCE to add ALL the sections.

While technically we should call llvm-objcopy for each section according
to the doc?

"--add-section <section=file>
Add a section named <section> with the contents of <file> to the output.
For ELF objects the section will be of type SHT_NOTE, if the name starts
with “.note”. Otherwise, it will have type SHT_PROGBITS. **Can be
specified multiple times to add multiple sections."**

This is however common to upstream, so workaround it first by disabling
the verification while we follow up to fix the problem.