[SYCL] Change the ext_intel_device_info spec to throw a feature not supported error when a query is not supported (#14576)
When an aspect is not supported by a device, the `ext_intel_device_info`
spec says to throw an invalid object runtime error when calling
`get_info` with the corresponding descriptor. This error description is
both vague and not very descriptive of the actual thing that went wrong,
namely that the query is simply not supported by a device. This PR
changes that to instead say to throw a feature not supported runtime
error.

---------

Co-authored-by: Greg Lueck <gregory.m.lueck@intel.com>