[SYCL] Remove unnamed lambda check from headers (#14614)
Currently the SYCL headers checks for unnamed lambdas in the case where
they have been disabled. However, this falsely flags functor objects
which should not be covered by the unnamed lambda restriction. Luckily,
the compiler already issues diagnostics for illegal unnamed lambdas, so
this patch removes the check from the headers and instead relies on the
frontend for issuing diagnostics for the illegal cases.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>