[SYCL] Follow up fixes for group_sort extension (#14591)
* Change `memory_required` queries to return the exact size of local
memory which needs to be allocated for `group_sort` algorithm, currently
it is a bit more than required.
* Fix a bug in `align_key_value_scratch` where I didn't take into
account that `std::align` changes value of `KeysScratchSpace`. To find
scratch of the values I need to use original value of this variable.
* Fix mistake in the test where keys/values were mixed up. This is more
of a cosmetic change.
RunSortKeyValueOverGroup functions accepts two vectors - the first is
considered keys, and the second is considered values.
Verification is done in the same function (with the same assumption that
the first vector is keys and the second is values).
So it doesn't actually matter in which order aforementioned arguments
are provided, test still serves it purpose.