[SYCL][ABI-Break] Hide `getPlugin` symbol (#14616)
This function is not called anywhere from SYCL headers and therefore
there is no reason to have it exported.

It was made exported as part of ESIMD emulator plugin addition
(intel/llvm#2963), but that plugin was since
then deprecated and removed (intel/llvm#13295).

This is patch is essentially a by-product of intel/llvm#14145 and it is
done to simplify that change, i.e. PI plugins removal should not be
ABI-breaking by itself, we just need to cleanup some of our exported
symbols.