[SYCL][ABI-Break] Drop legacy overload of `queue::mem_advice` (#14618)
We made a mistake several years ago by introducing `pi_mem_advice`
argument into a public interface in intel/llvm#1239.

This mistake was caught a year later and addressed in intel/llvm#4100,
but unfortunately instead of removing this incorrect overload, it was
only deprecated. The deprecation wasn't done quite correctly, because
this method is *not* deprecated in SYCL 2020 - it *doesn't exists* in
SYCL 2020.

This PR removes the method completely, even though our messaging was
wrong. The assumption is that there are no actual users who went and
used undocumented non-public `pi_mem_advice` in their codebases.

This is patch is essentially a by-product of intel/llvm#14145 and it is
done to simplify that change, i.e. PI plugins removal should not be
ABI-breaking by itself, we just need to cleanup some of our exported
symbols.