[SYCL][ABI-Break] Hide some exported symbols (#14615)
Both `get_pi_error` and `set_pi_error` are only called from the library
and therefore there is no need for them to be exported.

This patch is essentially a by-product of #14145 and it is done to
simplify that change, i.e. PI plugins removal should not be ABI-breaking
by itself, we just need to cleanup some of our exported symbols.