[SYCL][ABI-Break] Drop `pi::assertion` (#14624)
This function should not be exported and looking at its uses it seems
like they call can be replaced with regular `assert` or a
`static_assert` without hurting any functionality.

This is patch is essentially a by-product of intel/llvm#14145 and it is
done to simplify that change, i.e. PI plugins removal should not be
ABI-breaking by itself, we just need to cleanup some of our exported
symbols.