[SYCL] Make `vec` conversion operator to scalar non-template
Template vs. non-template conversion operator behave differently for
```
  sycl::vec<int, 1> v;
  std::ignore = static_cast<bool>(v);
```

where non-template path selects the conversion operator and then uses
implicit `int->bool` conversion while the example fails to compile when
operator is a template.

This is important because the spec requires `vec::x()` and other
swizzles to return a swizzle class and not a `const DataT &` that it
currently does. However, if we'd fix the swizzle without addressing this
issues first the code like
```
  sycl::vec<int, 8> v = {...};
  if (v.x() == 42)
    ...
```
would fail to compile.