[SYCL] Fix enqueue functions taking both kernel and properties
The current implementation of the enqueue free functions taking both a
launch_config and a kernel do not properly process the properties. This
commit addresses this and adds a static assert about the properties
passed to these only applying to the launch of the kernel and not how
the compiler handles compiling the kernel.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>