[SYCL] Introduce new properties for USM allocations
This is an RFC for a set of new properties serving as hints to USM
allocation functions.

The current SYCL 2020 spec declares all allocation functions as
accepting a property list argument, but so far the semantics of these
values remain unspecified. We take advantage of it here by specifying a
set of properties that can be optionally used by the implementation to
inform the underlying allocator of specific use patterns. It's in many
ways analogous to posix's `madvise` and the extant SYCL madvise
extension, but since the flags are specified at allocation, it's
potentially more convenient and efficient for the underlying backend to
effectively utilize.

In addition to the flags already supported by most madvise
implementations we also introduce several flags useful on accelerators
that for some workloads may enable significant performance benefits.
e.g. `device_cache_noncoherent` has been seen to enable a doubling of
performance in some workloads with `malloc_host`, where the accelerator
and host don't need to agree on the contents of an allocation until
after a `.wait()` - a very common idiom. This extension was developed
for the Quicksilver nuclear simulation benchmark which significantly
benefits from the `device_cache_non_coherent` flag mentioned above to a
tune of ~160%/

Documentation is in
doc/extensions/proposed/sycl_ext_usm_properties.asciidoc.