[SYCL][E2E] Only grep if compilation is supported in KernelCompiler/sycl_device_flags.cpp (#14760)
After my last fix, another problem appeared (but only in post-commit).
This is a blind fix for that. Let me see if I can run that post-commit
pass from GitHub Actions.

The premature exit if compilation is not available was leading to test
failures once the test tried to call grep. I'm moving that check into
the test app itself, rather than use a `// RUN: ` directive.

Ultimately, we may want something like `// REQUIRE:
kernel_compiler::lang::sycl` but this is not the time/place for that.