[SYCL][Driver] Allow for -save-temps to work with -fsycl-host-compiler (#14751)
When using -save-temps, the driver will break down each compilation step
to multiple steps (preprocess, ir, asm, obj) for a good gathering of all
of the files needed for compilation. When performing offload
compilations, this behavior is valid with the exception of using a third
party compiler with -fsycl-host-compiler. The reason it does not work is
due to the fact that the third party compiler may or may not support
these separate step creations, so we error out letting the user know.

Update the behavior of the driver to collapse all of these compilation
steps into one so only the object creation step is performed when doing
the third party compilation.