[SYCL][NFC] Avoid warnings in SYCLSpecConstMaterializer and program_manager (#14770)
This commit avoids the following warnings in post-commit testing:
* SYCLSpecConstMaterializer had a loop purely for debug printing. If
`NDEBUG` is defined, the loop variable is unused, which causes a
warning. To address this, the use of `LLVM_DEBUG` is moved to outside
the full debug printing body.
* In program_manager, debug printing now checks whether debug printing
is enabled using `if constexpr`. This now means that `getFormatStr` is
unused, which in turn causes a warning. This commit marks it as
potentially unused to avoid this warning.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>