[SYCL] Add compile target to device image properties (#14757)
Previously, when compiling with any `intel_gpu_*` target, the device
image only had the generic `spir64_gen` as its value for the
[`DeviceTargetSpec`](https://github.com/intel/llvm/blob/450683b6fa1d1be1b9391905f43073b7a9555aa1/sycl/include/sycl/detail/pi.h#L1143)
in its device image, so there was no way to differentiate device images
compiled for a specific GPU in the runtime. Now, a new property is added
to the image specifying the specific `intel_gpu_*` value it is compiled
for.