[SYCL][NFC] Detach compiler interfaces from PI (#14700)
Our compiler interfaces do not depend on runtime bits and therefore they
shouldn't be tied to PI, UR or anything else. This PR is a by-product of
#14145 intended to refactor the code in advance so once we move to UR
there are less changes required.

Moreover, compiler interfaces are not used anywhere in public SYCL
headers and therefore they were moved into library.

As you can see, there are still some `pi` naming present in filenames
and namespaces, but this patch is huge enough already, so further
changes will be done as separate PRs.