[SYCL] remove `alignedAllocHost`'s `Kind` param
`alignedAllocHost` was introduced as part of what was described as a
bugfix in 01869a033a76 wherin `alignedAlloc` was simply renamed to
`alignedAllocHost`, but the `Kind` parameter (host|device|shared) was
left unchanged, and the bugfix actually appears to be in the calling
code. I'm not quite sure what happened there, but in any case we
shouldn't have a `Kind` parameter if we only ever intend to work on the
host (indeed the existing code treats Kind == host as an invariant,
returning an error otherwise).

It looks like a couple of other people tried to work around this when
adding other features (e.g. in 7df39237b30) but didn't fix this due to
ABI breaks.

Thus, this is a putatively NFC cleanup, but it's a change to the SYCL
runtime ABI so not "genuine-NFC".