[SYCL][Graph] Fix E2E disabled RUN lines
We were using the `x` prefix to the `RUN` lit command to comment out a `RUN`
check. However, LIT still detects this as a `RUN` command, which illustrated
in a CI fail of this test recently https://github.com/intel/llvm/actions/runs/10198164930/job/28215254220

Fixed by changing `xRUN` to `RUNx` which I've verified locally is not detected as a
`RUN` command by LIT, and can be used to comment the `RUN` command out.

Relates to #14473