[SYCL] Catch exceptions thrown in destructors (#14808)
In the same vein as #14273, this PR prevents exceptions from leaking in
additional destructors caught by Coverity.

I'd like to draw attention to `device_impl.cpp` however: A comment was
left suggesting that exceptions in the `device_impl` destructor be added
to the asynchronous exceptions list. Given that devices are usually
destroyed during shutdown, adding exceptions to the exceptions list
doesn't seem to make sense, as there would be nothing to handle the
exceptions anyway.

However, if this understanding is incorrect, and I should still add
exceptions to an asynchronous exceptions list, please let me know.
Thanks!