[CUDA] Don't import XPTI symbols in the plugin library
The CUDA plugin builds an XPTI file directly.
By default the symbol visibility in that XPTI file is presumed
to import symbols, but there are no XPTI symbols being exported,
since XPTI is not built as a separate library.

This causes a compilation failure on Windows.
The fix is to define `XPTI_STATIC_LIBRARY`,
which changes the visibility of symbols -
on Windows this means no longer using `dllimport`
(and neither using `dllexport`).