[NVPTX][AMDGPU] Move annotation creation out of clang (#14634)
This patch refactors the way we lower SYCL attributes and properties to
NVVM annotations, through function attributes and metadata. It unifies
the flow better with the SPIR-V paths at the same time.

Previously we had:

1. Clang handling function attributes in two places:
1. CodeGenFunction did generic lowering of attributes to function
metadata
2. NVPTXTargetCodeGenInfo did its own additional lowering of attributes
to NVVM annotations
2. Kernel properties being handled in clang. NVPTXTargetCodeGenInfo
lowered kernel properties, which had already been converted to function
attributes, to NVVM annotations.
3. Kernel properties for HIP/CUDA *not* being turned into function
metadata. Because function metadata is how the ComputeModuleRuntimeInfo
library creates runtime info, this meant that target backends couldn't
act on properties, as they were lost during lowering.

Fundamentally, clang is not the correct place for handling SYCL kernel
properties as it unnecessarily touches the front-end for what is really
library code.

With this patch, we now have:

1. Clang handles function attributes in one place, lowering to function
metadata
2. SYCLLowerIR lowers kernel properties to the same function metadata:
now working for HIP and CUDA kernels too. Because function metadata is
consistently present for kernel properties, they behave as expected on
HIP/CUDA targets now.
3. A new pass converts function metadata to NVVM annotations, replacing
the code we've taken out of clang.