[SYCL][UR] Use static cast instead of ur::cast (#14830)
UR cast is using reinterpret cast under the hood, which was failing for
an
unsigned long being cast to an unsigned long long. From the C++ spec:

> An expression of integral, enumeration, pointer, or pointer-to-member 
type can be converted to its own type. The resulting value is the same
as
the value of expression.

Since `unsigned long` is not the same type as `unsigned long long`, the
`reinterpret_cast` fails.

Ping @aarongreig