[SYCL] Filter out the header and footer from the dependency output. (#14933)
`Cmake` uses dependency information generated by the use of the` -MD`
option to decide if a file needs to be rebuilt. Currently the header and
footer are not filtered out from the dependency output. This leads to
the issue that `Cmake` thinks the source file needs to always be
rebuilt.
This is a patch to fix the issue.
See https://github.com/intel/llvm/discussions/14798 for more details.