[SYCL] Add work-around for event leak in profiling tag (#14985)
Due to a bug in the L0 UR adapter, the profiling tag extension leaks UR
events on out-of-order queues. This is not due to the profiling tag
events themselves, but rather due to the need for a barrier enforcing
correct ordering of the inserted tag. Since the barrier ensures
completion prior to the profiling tag executing, the output event is not
needed, but the L0 adapter leaks the event if no output event is
specified. To combat this, this work-around passes an output event and
immediately frees it after the barrier has been submitted.

See https://github.com/oneapi-src/unified-runtime/issues/1947.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>