[SYCL][E2E][Bindless] Check and use DedicatedAllocation in vulkan interop test for L0 (#14967)
Vulkan interop in L0 backend currently has an issue that image sizes are
not correctly passed to graphics memory management (GMM). Using
dedicated allocation is an approach to solve the issue. In this case,
image and device memory has 1:1 mapping and GMM is aware of the image
info.
This PR enables VK_KHR_GET_MEMORY_REQUIREMENTS_2_EXTENSION_NAME
extension and checks if
VkMemoryDedicatedRequirements.requiresDedicatedAllocation is true. The
value is false by default in CUDA and L0 backends.