[SYCL] Drop UrArray from unit-tests
`UrArray` is a wrapper intended to extend lifetime of mock objects so
that native objects referncing mock data can be accessed by SYCL RT
without any issues.

This commit removes the wrapper to simplify writing unit-tests:
essentially there is now one less custom class to care about,
collections of properties and entry points are now handled using regular
`std::vector` and initializer lists.

The cost of this is a slightly more complicated implementation for
`UrPropertySet` and `UrImage` that now need to replicate functionality
of that extra storage from `UrArray`.