[SYCL][CUDA][HIP] Define backend macros based on compiler
Currently the backend macros are only defined when the runtime adapters
for the given backend are built.

It's more accurate to use whether or not the compiler can target these
backends, regardless of whether the runtime elements are present.

It should be possible to compile an application for a backend without
having to install the matching runtime.