Change urDeviceCreateWithNativeHandle to take an adapter handle.
It currently takes a platform handle, which is problematic for the sycl
RT because its make_device api only takes a native handle, so to figure
out the correct platform handle to pass at best we'd need to do some
backend specific querying of the native object, but even then that isn't
always possible as not all backends have a platform equivalent.

The platform handle does currently enable an optional (slightly) faster
path to return the correct device in some adapter implementations but
this isn't essential for them to work, so really its primary purpose is
to serve as the wrapped UR handle for the loader to work. This purpose
is equally well served by an adapter handle, which will also be a lot
easier for the sycl rt to correctly provide.