[SYCL] Add AMDGPU_kernel calling convention to detected kernels (#14581)
`free_function_kernels.cpp` test had a bug where the kernels with
demangled name for free functions will be deleted in the
`sycl-post-link` step of compilation. And this happened as AMD kernels
was not detected due to a missing condition. This was not detected
before as the HIP device on CI doesn't have `usm_shared_allocations`
aspect available so it was detected as unsupported but when I tried it
locally with a device with `usm_shared_allocations` aspect available,
the test was failing.