[SPIR-V] Add W/A to finish SROA for cooperative matrices
SROA has troubles processing structures containing TargetExtType,
which in some cases will lead to a situation, when OpAccessChain would
attempt to access a structure containing a cooperative matrix, but with
indexes set to access the matrix itself, which is invalid. This PR adds
routine that finds alloca to such structure and replaces it with alloca
to cooperative matrix type.

Signed-off-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>