[SYCL] WA for stream events cross dependency elimination (#14797)
Current stream implementation has a problem with cross dependency for
inorder queue case. Both stream commands: kernel which writes and host
task which prints are submitted to the same queue now, inorder queue
mechanism of synchronization will add explicit dependency on kernel to
the host task (this approach helps up to provide synchronization for
kernel vs host tasks since no native sync is supported). Although kernel
will add flush host task to its post-complete events storage to be able
to wait for its completion in its own wait. This brings cross dependency
that prevents event release.
Current WA stores host task using weak reference that works in the same
way but prevents cross dependency. Host task doesn't support (and not
planned to support) post enqueue cleanup that means that its Command
instance could be deleted only after host task completion and later.
That means that absence of object in stored weak pointer is a solid sign
of host task completion and we have no need to wait on it. If it is
present we will try to wait on it if it is requested.
This should be a temporal solution. Solution to be implemented later is
elimination of forward references for stream and implementing a kind of
aggregated events.

---------

Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>
Co-authored-by: Sergey Semenov <sergey.semenov@intel.com>