[SYCL][Docs] Allow copy-construction of device_global
This commit makes it possible to copy-construct device_global variables
if they do not have the device_image_scope property. The restriction on
device_image_scope is due to static construction not being allowed in
device code, which they would require, while other device_globals have
USM storage which will be initialized by the host code, so the
constructor on the device is a simple zero-initialization.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>