Fix any mux abi calls to be abi conformant
Native cpu can make calls to mux builtins such as shuffle which are
ABI compliant but are not what is expected by ock passes. This fixes
them up by making the created function headers a wrapper which calls
the correct mux functions.

This currently only handle a small number of cases for shuffle such
as when a vector i2 is replaced with double or byval is used. It will
be expanded over time.