Add dlerror output to loader logs
Failing to load an adapter is not an error in most cases the adapter is
just not present.

However there's a lot of other reasons why an adapter may fail to load,
such as a missing driver library or an unsupported driver. In these
cases `dlerror` output is crucial to troubleshoot the issue.