[SYCL][E2E] Remove leaking part of non-uniform-wk-gp-test
As shown in https://github.com/intel/llvm/issues/15111, currently
sycl::stream will leak resources if an exception is thrown during launch
of its associated command. This was in turn causing
sycl/test-e2e/Basic/diagnostics/non-uniform-wk-gp-test.cpp to leak
resources simply because it was using the stream in its testing.
However, since the kernel should never execute, the stream did not
serve a fundamental function in the test and was only there for
additional debugging information in the case where the test would fail.
As such, it is arguably better to improve coverage than to disable cases
because of some potentially helpful additional output in the failing
case.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>